+++ changelog.md
@@ -1,3 +1,7 @@
+# 0.3.2.0
+- Changed string parsing; parsing of escaped strings is now very vast
+- Removed bytestring parser
+
 # 0.3.0.4
 - Fixed bug in safestring
 - Fixed test so it doesn't depend on versions of other packages